# be/src/api-spec/public/api.yaml
openapi: 3.0.0

info:
  title: Library Management System (LMS) API Specification
  version: 1.0.0
  description: |
    REST API for the LMS platform.
    Handles user, book, borrow, post, comment management.
  contact:
    name: API Support
    email: baonamdao05@gmail.com

servers:
  - url: http://localhost:8080/api
    description: Local development

tags:
  - name: User
    description: User entity
  - name: Book
    description: Book entity
  - name: Borrow
    description: Borrow entity
  - name: Post
    description: Post entity
  - name: Comment
    description: Comment entity

# ------- PATHS -------

paths:
  /users:
    get:
      tags: [User]
      summary: 'To get all users'
      operationId: 'findAll'
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirParam'
        - name: username
          required: false
          in: query
          description: Find by username
          schema:
            type: string
        - name: fullName
          required: false
          in: query
          description: Find by full name
          schema:
            type: string
        - name: role
          required: false
          in: query
          description: Find by role
          schema:
            type: string
      
      responses:
        '200':
          description: All users are found and returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/General Response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User Response'
                example:
                  code: "200"
                  message: "Success"
                  data:
                    - id: 1
                      username: "tyler"
                      fullName: "Tyler Dao"
                      role: "ROLE_USER"
                      phoneNumber: "0123456789"
                      address: "Hanoi"
                      email: "tyler@gmail.com"
                    - id: 2
                      username: "tyler2"
                      fullName: "Tyler Dao 2"
                      role: "ROLE_ADMIN"
                      phoneNumber: "0123456789"
                      address: "Hanoi"
                      email: "tyler2@gmail.com"
        '401':
          $ref: '#/components/responses/AuthException'
        '403':
          $ref: '#/components/responses/AccessDeniedException'
        '500':
          $ref: '#/components/responses/RunTimeException'
  /user:
    post:
      tags: [User]
      summary: Create new user
      operationId: saveUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User Input'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/General Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User Response'
        '401':
          $ref: '#/components/responses/AuthException'
        '403':
          $ref: '#/components/responses/AccessDeniedException'
        '500':
          $ref: '#/components/responses/RunTimeException'

# ------- REUSABLE COMPONENTS -------
components:

  # Auth Scheme (Moved safely inside components root)
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Schemas
  schemas:
    Register Input:
      type: object
      required: [username, password, phoneNumber, role, fullName]
      properties:
        username:
          type: string
        role:
          type: string
        phoneNumber:
          type: string
        fullName:
          type: string
        address:
          type: string
        email:
          type: string
          format: email

    Login Input:
      type: object
      required: [email, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    Login Response:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        role:
          type: string

    User Input:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
          nullable: true
        email:
          type: string
          nullable: true

    User Response:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
        role:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
          nullable: true
        email:
          type: string
          nullable: true

    Role Response:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          nullable: false

    General Response:
      type: object
      properties:
        code:
          type: string
          example: "500"
        message:
          type: string
          example: "Internal server error"
        data:
          type: object
          nullable: true

  # Reusable Parameters
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10

    SortByParam:
      name: sortBy
      description: Sort by a field of object
      in: query
      schema:
        type: string
        default: "id"

    SortDirParam:
      name: sortDir
      in: query
      schema:
        type: string
        default: "asc"
        enum:
          - "asc"
          - "desc"

  # Reusable Responses
  responses:
    RunTimeException:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/General Response'
          example:
            code: "500"
            message: "Internal server error"
            data: null

    AuthException:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/General Response'
          example:
            code: "401"
            message: "Unauthorized API request"
            data: null

    AccessDeniedException:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/General Response'
          example:
            code: "403"
            message: "Access denied"
            data: null